zstd - refactoring

+++ src/main/java/ru/olegcherednik/zip4jvm/io/zstd/BlockCompressionState.java
@@ -20,12 +20,12 @@ class BlockCompressionState
     public final int[] hashTable;
     public final int[] chainTable;
 
-    private final long baseAddress;
+    private final int baseAddress;
 
     // starting point of the window with respect to baseAddress
     private int windowBaseOffset;
 
-    public BlockCompressionState(CompressionParameters parameters, long baseAddress)
+    public BlockCompressionState(CompressionParameters parameters, int baseAddress)
     {
         this.baseAddress = baseAddress;
         hashTable = new int[1 << parameters.getHashLog()];
@@ -48,7 +48,7 @@ class BlockCompressionState
         }
     }
 
-    public long getBaseAddress()
+    public int getBaseAddress()
     {
         return baseAddress;
     }
