+++ src/main/java/ru/olegcherednik/zip4jvm/io/zstd/XxHash64.java
@@ -26,7 +26,7 @@ final class XxHash64 {
 
     private XxHash64() {}
 
-    public static long hash(long seed, byte[] inputBase, long address, int length) {
+    public static long hash(long seed, byte[] inputBase, int address, int length) {
         long hash;
         if (length >= 32) {
             hash = updateBody(seed, inputBase, address, length);
@@ -43,7 +43,7 @@ final class XxHash64 {
         return updateTail(hash, inputBase, address, index, length);
     }
 
-    private static long updateTail(long hash, byte[] inputBase, long address, int index, int length) {
+    private static long updateTail(long hash, byte[] inputBase, int address, int index, int length) {
         while (index <= length - 8) {
             hash = updateTail(hash, UnsafeUtil.getLong(inputBase, address + index));
             index += 8;
@@ -64,7 +64,7 @@ final class XxHash64 {
         return hash;
     }
 
-    private static long updateBody(long seed, byte[] inputBase, long address, int length) {
+    private static long updateBody(long seed, byte[] inputBase, int address, int length) {
         long v1 = seed + PRIME64_1 + PRIME64_2;
         long v2 = seed + PRIME64_2;
         long v3 = seed;
